Modified ocean.py and balancer to work with 32 bit floats instead of 64 bit ones. Fixed bug in Doppler calculation in skim_raw.

+++ oceansar/surfaces/balancer.py
@@ -81,43 +81,43 @@ class OceanSurfaceBalancer(object):
 
         # Process-dependent properties
         self.Ny = np.int(self.counts[self.rank] / self.Nx)
-        self.y = np.empty(self.Ny)
+        self.y = np.empty(self.Ny, dtype=np.float32)
         if self.rank == self.root:
             y = (np.ascontiguousarray(surface.y),
-                 (self.counts / self.Nx, self.displ / self.Nx), MPI.DOUBLE)
+                 (self.counts / self.Nx, self.displ / self.Nx), MPI.FLOAT)
         else:
             y = None
-        self.comm.Scatterv(y, (self.y, MPI.DOUBLE), root=self.root)
+        self.comm.Scatterv(y, (self.y, MPI.FLOAT), root=self.root)
 
         # INITIALIZE SURFACE
         # Memory allocation (LOW (0) / HIGH (1) dt values)
         if 'D' in self.compute:
-            self._Dx = np.empty(
-                2 * int(self.counts[self.rank])).reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx))
-            self._Dy = np.empty_like(self._Dx)
-            self._Dz = np.empty_like(self._Dx)
+            self._Dx = (np.empty(2 * int(self.counts[self.rank]), dtype=np.float32).
+                        reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx)))
+            self._Dy = np.empty_like(self._Dx, dtype=np.float32)
+            self._Dz = np.empty_like(self._Dx, dtype=np.float32)
         if 'Diff' in self.compute:
-            self._Diffx = np.empty(
-                2 * int(self.counts[self.rank])).reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx))
+            self._Diffx = (np.empty(2 * int(self.counts[self.rank]), dtype=np.float32).
+                           reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx)))
             self._Diffy = np.empty_like(self._Diffx)
         if 'Diff2' in self.compute:
-            self._Diffxx = np.empty(
-                2 * int(self.counts[self.rank])).reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx))
+            self._Diffxx = (np.empty(2 * int(self.counts[self.rank]), dtype=np.float32).
+                            reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx)))
             self._Diffyy = np.empty_like(self._Diffxx)
             self._Diffxy = np.empty_like(self._Diffxx)
         if 'V' in self.compute:
-            self._Vx = np.empty(
-                2 * int(self.counts[self.rank])).reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx))
+            self._Vx = (np.empty(2 * int(self.counts[self.rank]), dtype=np.float32).
+                        reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx)))
             self._Vy = np.empty_like(self._Vx)
             self._Vz = np.empty_like(self._Vx)
         if 'A' in self.compute:
-            self._Ax = np.empty(
-                2 * int(self.counts[self.rank])).reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx))
+            self._Ax = (np.empty(2 * int(self.counts[self.rank]), dtype=np.float32).
+                        reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx)))
             self._Ay = np.empty_like(self._Ax)
             self._Az = np.empty_like(self._Ax)
         if 'hMTF' in self.compute:
-            self._hMTF = np.empty(
-                2 * int(self.counts[self.rank])).reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx))
+            self._hMTF = (np.empty(2 * int(self.counts[self.rank]), dtype=np.float32).
+                          reshape(2, int(self.counts[self.rank] / self.Nx), int(self.Nx)))
 
         self.dt = dt
         self.t_l_last = -1.
@@ -131,10 +131,10 @@ class OceanSurfaceBalancer(object):
     @t.setter
     def t(self, value):
 
-        self._t = value
+        self._t = np.float32(value)
 
         # Update low/high times
-        t_l = np.floor(self._t / self.dt) * self.dt
+        t_l = np.float32(np.floor(self._t / self.dt) * self.dt)
         t_h = t_l + self.dt
 
         if (t_l != self.t_l_last) or (t_h != self.t_h_last):
@@ -172,40 +172,40 @@ class OceanSurfaceBalancer(object):
                     self.surface.t = t_i[1]
                     if 'D' in self.compute:
                         Dx_f = (self.surface.Dx,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                         Dy_f = (self.surface.Dy,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                         Dz_f = (self.surface.Dz,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                     if 'Diff' in self.compute:
                         Diffx_f = (self.surface.Diffx,
-                                   (self.counts, self.displ), MPI.DOUBLE)
+                                   (self.counts, self.displ), MPI.FLOAT)
                         Diffy_f = (self.surface.Diffy,
-                                   (self.counts, self.displ), MPI.DOUBLE)
+                                   (self.counts, self.displ), MPI.FLOAT)
                     if 'Diff2' in self.compute:
                         Diffxx_f = (self.surface.Diffxx,
-                                    (self.counts, self.displ), MPI.DOUBLE)
+                                    (self.counts, self.displ), MPI.FLOAT)
                         Diffyy_f = (self.surface.Diffyy,
-                                    (self.counts, self.displ), MPI.DOUBLE)
+                                    (self.counts, self.displ), MPI.FLOAT)
                         Diffxy_f = (self.surface.Diffxy,
-                                    (self.counts, self.displ), MPI.DOUBLE)
+                                    (self.counts, self.displ), MPI.FLOAT)
                     if 'V' in self.compute:
                         Vx_f = (self.surface.Vx,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                         Vy_f = (self.surface.Vy,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                         Vz_f = (self.surface.Vz,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                     if 'A' in self.compute:
                         Ax_f = (self.surface.Ax,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                         Ay_f = (self.surface.Ay,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                         Az_f = (self.surface.Az,
-                                (self.counts, self.displ), MPI.DOUBLE)
+                                (self.counts, self.displ), MPI.FLOAT)
                     if 'hMTF' in self.compute:
                         hMTF_f = (self.surface.hMTF,
-                                  (self.counts, self.displ), MPI.DOUBLE)
+                                  (self.counts, self.displ), MPI.FLOAT)
 
                 else:
                     if 'D' in self.compute:
@@ -232,47 +232,47 @@ class OceanSurfaceBalancer(object):
 
                 if 'D' in self.compute:
                     self.comm.Scatterv(
-                        Dx_f, (self._Dx[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Dx_f, (self._Dx[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Dy_f, (self._Dy[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Dy_f, (self._Dy[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Dz_f, (self._Dz[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Dz_f, (self._Dz[int(t_i[0])], MPI.FLOAT), root=self.root)
                 if 'Diff' in self.compute:
                     self.comm.Scatterv(
-                        Diffx_f, (self._Diffx[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Diffx_f, (self._Diffx[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Diffy_f, (self._Diffy[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Diffy_f, (self._Diffy[int(t_i[0])], MPI.FLOAT), root=self.root)
                 if 'Diff2' in self.compute:
                     self.comm.Scatterv(
-                        Diffxx_f, (self._Diffxx[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Diffxx_f, (self._Diffxx[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Diffyy_f, (self._Diffyy[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Diffyy_f, (self._Diffyy[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Diffxy_f, (self._Diffxy[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Diffxy_f, (self._Diffxy[int(t_i[0])], MPI.FLOAT), root=self.root)
                 if 'V' in self.compute:
                     self.comm.Scatterv(
-                        Vx_f, (self._Vx[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Vx_f, (self._Vx[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Vy_f, (self._Vy[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Vy_f, (self._Vy[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Vz_f, (self._Vz[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Vz_f, (self._Vz[int(t_i[0])], MPI.FLOAT), root=self.root)
                 if 'A' in self.compute:
                     self.comm.Scatterv(
-                        Ax_f, (self._Ax[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Ax_f, (self._Ax[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Ay_f, (self._Ay[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Ay_f, (self._Ay[int(t_i[0])], MPI.FLOAT), root=self.root)
                     self.comm.Scatterv(
-                        Az_f, (self._Az[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        Az_f, (self._Az[int(t_i[0])], MPI.FLOAT), root=self.root)
                 if 'hMTF' in self.compute:
                     self.comm.Scatterv(
-                        hMTF_f, (self._hMTF[int(t_i[0])], MPI.DOUBLE), root=self.root)
+                        hMTF_f, (self._hMTF[int(t_i[0])], MPI.FLOAT), root=self.root)
 
         self.t_l_last = t_l
         self.t_h_last = t_h
 
         # Apply linear interpolation
-        w_h = (self._t - t_l) / self.dt
-        w_l = 1. - w_h
+        w_h = np.float32((self._t - t_l) / self.dt)
+        w_l = np.float32(1. - w_h)
 
         if 'D' in self.compute:
             self.Dx = w_l * self._Dx[0] + w_h * self._Dx[1]
